Download a file from the ToL SFTP server.

Download an alignment, sequence, or tree file from the ToL SFTP server.

download_tol(download_link = NULL, save_dir = NULL)

Arguments

download_link

A string specifying the URL to download the file from. You can get a download URL for a particular specimen or gene using lookup_tol().

save_dir

A string specifying the folder to save the download in. If no value is provided, here will be used.

Details

The Tree of Life is a database of specimens sequenced as part of Kew's efforts to build a comprehensive evolutionary tree of life for flowering plants.

Sequence, alignment, and Newick tree files are help on an SFTP server for download. The URLs to access these are stored in entries for specimens and genes in the ToL database. These can be accessed by either using search_tol() to get all specimens for a particular order, family, genus, or species or by looking up a specific specimen or gene using lookup_tol()

References

Baker W.J., Bailey P., Barber V., Barker A., Bellot S., Bishop D., Botigue L.R., Brewer G., Carruthers T., Clarkson J.J., Cook J., Cowan R.S., Dodsworth S., Epitawalage N., Francoso E., Gallego B., Johnson M., Kim J.T., Leempoel K., Maurin O., McGinnie C., Pokorny L., Roy S., Stone M., Toledo E., Wickett N.J., Zuntini A.R., Eiserhardt W.L., Kersey P.J., Leitch I.J. & Forest F. 2021. A Comprehensive Phylogenomic Platform for Exploring the Angiosperm Tree of Life. Systematic Biology, 2021; syab035, https://doi.org/10.1093/sysbio/syab035

See also

Examples

if (FALSE) {
 # download a specimen fasta file
 specimen_info <- lookup_tol("1296")
 download_tol(specimen_info$fasta_file_url)

 # download a gene alignment file
 gene_info <- lookup_tol("51", type="gene")
 download_tol(gene_info$alignment_file_url)

 # download the gene tree
 download_tol(gene_info$tree_file_url)
}