Get citation for Kew data resource.

Given the result of a query to a Kew data resource, get the appropriate citation.

kew_citation(x)

Arguments

x

Result of a call to search_powo(), lookup_powo(), search_wcvp(), lookup_wcvp(), search_ipni(), lookup_ipni(), search_tol(), load_tol(), match_knms(), or match_krs()

Value

A citation object with a print method for nice display.

Examples

r <- search_powo(list(characteristic="poison"))
kew_citation(r)
#> 
#> To cite POWO in publication please use:
#> 
#>   POWO (2021). Plants of the World Online. Facilitated by the Royal
#>   Botanic Gardens, Kew. http://www.plantsoftheworldonline.org/.
#>   Accessed 2021-12-17
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {Plants of the World Online.},
#>     author = {{POWO}},
#>     year = {2021},
#>     note = {Facilitated by the Royal Botanic Gardens, Kew.},
#>     accessed = {2021-12-17},
#>     url = {http://www.plantsoftheworldonline.org/},
#>   }
#> 

r <- lookup_powo("320035-2")
kew_citation(r)
#> 
#> To cite POWO in publication please use:
#> 
#>   POWO (2021). Plants of the World Online. Facilitated by the Royal
#>   Botanic Gardens, Kew. http://www.plantsoftheworldonline.org/.
#>   Accessed 2021-12-17
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {Plants of the World Online.},
#>     author = {{POWO}},
#>     year = {2021},
#>     note = {Facilitated by the Royal Botanic Gardens, Kew.},
#>     accessed = {2021-12-17},
#>     url = {http://www.plantsoftheworldonline.org/},
#>   }
#> 

r <- search_wcvp(list(genus="Poa"))
kew_citation(r)
#> 
#> To cite WCVP in publication please use:
#> 
#>   WCVP (2021). World Checklist of Vascular Plants, version 2.0.
#>   Facilitated by the Royal Botanic Gardens, Kew.
#>   http://wcvp.science.kew.org/. Accessed 2021-12-17
#> 
#>   Govaerts, R., Nic Lughadha, E., Black, N., Turner, R. and Paton, A.
#>   (2021). The World Checklist of Vascular Plants, a continuously
#>   updated resource for exploring global plant diversity. Scientific
#>   Data, 8(215), https://doi.org/10.1038/s41597-021-00997-6
#> 
#> To see these entries in BibTeX format, use 'print(<citation>,
#> bibtex=TRUE)', 'toBibtex(.)', or set
#> 'options(citation.bibtex.max=999)'.
#> 

r <- lookup_wcvp("320035-2")
kew_citation(r)
#> 
#> To cite WCVP in publication please use:
#> 
#>   WCVP (2021). World Checklist of Vascular Plants, version 2.0.
#>   Facilitated by the Royal Botanic Gardens, Kew.
#>   http://wcvp.science.kew.org/. Accessed 2021-12-17
#> 
#>   Govaerts, R., Nic Lughadha, E., Black, N., Turner, R. and Paton, A.
#>   (2021). The World Checklist of Vascular Plants, a continuously
#>   updated resource for exploring global plant diversity. Scientific
#>   Data, 8(215), https://doi.org/10.1038/s41597-021-00997-6
#> 
#> To see these entries in BibTeX format, use 'print(<citation>,
#> bibtex=TRUE)', 'toBibtex(.)', or set
#> 'options(citation.bibtex.max=999)'.
#> 

r <- search_ipni(list(publishing_author="L."))
#> No encoding supplied: defaulting to UTF-8.
kew_citation(r)
#> 
#> To cite IPNI in publication please use:
#> 
#>   IPNI (2021). International Plant Names Index. The Royal Botanic
#>   Gardens, Kew, Harvard University Herbaria & Libraries and Australian
#>   National Botanic Gardens. https://ipni.org/. Accessed 2021-12-17
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {International Plant Names Index},
#>     author = {{IPNI}},
#>     year = {2021},
#>     note = {The Royal Botanic Gardens, Kew, Harvard University Herbaria & Libraries and Australian National Botanic Gardens},
#>     accessed = {2021-12-17},
#>     url = {https://ipni.org/},
#>   }
#> 

r <- lookup_ipni("12653-1")
#> No encoding supplied: defaulting to UTF-8.
kew_citation(r)
#> 
#> To cite IPNI in publication please use:
#> 
#>   IPNI (2021). International Plant Names Index. The Royal Botanic
#>   Gardens, Kew, Harvard University Herbaria & Libraries and Australian
#>   National Botanic Gardens. https://ipni.org/. Accessed 2021-12-17
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {International Plant Names Index},
#>     author = {{IPNI}},
#>     year = {2021},
#>     note = {The Royal Botanic Gardens, Kew, Harvard University Herbaria & Libraries and Australian National Botanic Gardens},
#>     accessed = {2021-12-17},
#>     url = {https://ipni.org/},
#>   }
#> 

r <- search_tol("Poa")
#> No encoding supplied: defaulting to UTF-8.
kew_citation(r)
#> 
#> To cite ToL in publication please use:
#> 
#>   Baker W.J., Bailey P., Barber V., Barker A., Bellot S., Bishop D.,
#>   Botigue L.R., Brewer G., Carruthers T., Clarkson J.J., Cook J., Cowan
#>   R.S., Dodsworth S., Epitawalage N., Francoso E., Gallego B., Johnson
#>   M., Kim J.T., Leempoel K., Maurin O., McGinnie C., Pokorny L., Roy
#>   S., Stone M., Toledo E., Wickett N.J., Zuntini A.R., Eiserhardt W.L.,
#>   Kersey P.J., Leitch I.J. and Forest F. (2021). A Comprehensive
#>   Phylogenomic Platform for Exploring the Angiosperm Tree of Life.
#>   Systematic Biology, syab035, https://doi.org/10.1093/sysbio/syab035
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Article{,
#>     title = {A Comprehensive Phylogenomic Platform for Exploring the Angiosperm Tree of Life},
#>     author = {Baker W.J. and Bailey P. and Barber V. and Barker A. and Bellot S. and Bishop D. and Botigue L.R. and Brewer G. and Carruthers T. and Clarkson J.J. and Cook J. and Cowan R.S. and Dodsworth S. and Epitawalage N. and Francoso E. and Gallego B. and Johnson M. and Kim J.T. and Leempoel K. and Maurin O. and McGinnie C. and Pokorny L. and Roy S. and Stone M. and Toledo E. and Wickett N.J. and Zuntini A.R. and Eiserhardt W.L. and Kersey P.J. and Leitch I.J. and Forest F.},
#>     year = {2021},
#>     note = {The Royal Botanic Gardens, Kew, Harvard University Herbaria & Libraries and Australian National Botanic Gardens},
#>     journal = {Systematic Biology},
#>     volume = {syab035},
#>     url = {https://doi.org/10.1093/sysbio/syab035},
#>   }
#> 

r <- lookup_tol(2717)
#> No encoding supplied: defaulting to UTF-8.
kew_citation(r)
#> 
#> To cite ToL in publication please use:
#> 
#>   Baker W.J., Bailey P., Barber V., Barker A., Bellot S., Bishop D.,
#>   Botigue L.R., Brewer G., Carruthers T., Clarkson J.J., Cook J., Cowan
#>   R.S., Dodsworth S., Epitawalage N., Francoso E., Gallego B., Johnson
#>   M., Kim J.T., Leempoel K., Maurin O., McGinnie C., Pokorny L., Roy
#>   S., Stone M., Toledo E., Wickett N.J., Zuntini A.R., Eiserhardt W.L.,
#>   Kersey P.J., Leitch I.J. and Forest F. (2021). A Comprehensive
#>   Phylogenomic Platform for Exploring the Angiosperm Tree of Life.
#>   Systematic Biology, syab035, https://doi.org/10.1093/sysbio/syab035
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Article{,
#>     title = {A Comprehensive Phylogenomic Platform for Exploring the Angiosperm Tree of Life},
#>     author = {Baker W.J. and Bailey P. and Barber V. and Barker A. and Bellot S. and Bishop D. and Botigue L.R. and Brewer G. and Carruthers T. and Clarkson J.J. and Cook J. and Cowan R.S. and Dodsworth S. and Epitawalage N. and Francoso E. and Gallego B. and Johnson M. and Kim J.T. and Leempoel K. and Maurin O. and McGinnie C. and Pokorny L. and Roy S. and Stone M. and Toledo E. and Wickett N.J. and Zuntini A.R. and Eiserhardt W.L. and Kersey P.J. and Leitch I.J. and Forest F.},
#>     year = {2021},
#>     note = {The Royal Botanic Gardens, Kew, Harvard University Herbaria & Libraries and Australian National Botanic Gardens},
#>     journal = {Systematic Biology},
#>     volume = {syab035},
#>     url = {https://doi.org/10.1093/sysbio/syab035},
#>   }
#> 

tree <- load_tol()
#> No encoding supplied: defaulting to UTF-8.
kew_citation(tree)
#> 
#> To cite ToL in publication please use:
#> 
#>   Baker W.J., Bailey P., Barber V., Barker A., Bellot S., Bishop D.,
#>   Botigue L.R., Brewer G., Carruthers T., Clarkson J.J., Cook J., Cowan
#>   R.S., Dodsworth S., Epitawalage N., Francoso E., Gallego B., Johnson
#>   M., Kim J.T., Leempoel K., Maurin O., McGinnie C., Pokorny L., Roy
#>   S., Stone M., Toledo E., Wickett N.J., Zuntini A.R., Eiserhardt W.L.,
#>   Kersey P.J., Leitch I.J. and Forest F. (2021). A Comprehensive
#>   Phylogenomic Platform for Exploring the Angiosperm Tree of Life.
#>   Systematic Biology, syab035, https://doi.org/10.1093/sysbio/syab035
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Article{,
#>     title = {A Comprehensive Phylogenomic Platform for Exploring the Angiosperm Tree of Life},
#>     author = {Baker W.J. and Bailey P. and Barber V. and Barker A. and Bellot S. and Bishop D. and Botigue L.R. and Brewer G. and Carruthers T. and Clarkson J.J. and Cook J. and Cowan R.S. and Dodsworth S. and Epitawalage N. and Francoso E. and Gallego B. and Johnson M. and Kim J.T. and Leempoel K. and Maurin O. and McGinnie C. and Pokorny L. and Roy S. and Stone M. and Toledo E. and Wickett N.J. and Zuntini A.R. and Eiserhardt W.L. and Kersey P.J. and Leitch I.J. and Forest F.},
#>     year = {2021},
#>     note = {The Royal Botanic Gardens, Kew, Harvard University Herbaria & Libraries and Australian National Botanic Gardens},
#>     journal = {Systematic Biology},
#>     volume = {syab035},
#>     url = {https://doi.org/10.1093/sysbio/syab035},
#>   }
#> 

match <- match_knms("Poa annua")
kew_citation(match)
#> 
#> To cite KNMS in publication please use:
#> 
#>   KNMS (2021). Kew Names Matching Service.
#>   http://namematch.science.kew.org/. Accessed 2021-12-17
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {Kew Names Matching Service},
#>     author = {{KNMS}},
#>     year = {2021},
#>     accessed = {2021-12-17},
#>     url = {http://namematch.science.kew.org/},
#>   }
#> 

match <- match_krs("Poa annua")
kew_citation(match)
#> 
#> To cite KRS in publication please use:
#> 
#>   KRS (2016). Kew Reconciliation Service.
#>   http://data1.kew.org/reconciliation/. Accessed 2021-12-17
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {Kew Reconciliation Service},
#>     author = {{KRS}},
#>     year = {2016},
#>     accessed = {2021-12-17},
#>     url = {http://data1.kew.org/reconciliation/},
#>   }
#>