Look up a taxon in POWO.

Request the record for a taxon in Plants of the World Online (POWO) using the IPNI ID.

lookup_powo(taxonid, distribution = FALSE)

Arguments

taxonid
A string containing a valid IPNI ID.

distribution
Include distribution in results (default FALSE).

Value

A powo_taxon object, which is a simple structure with fields for each of the fields returned by the lookup API, as well as the the httr response object.

Details

Plants of the World Online (POWO) is a database of information on the world's flora. It curates information from published floras and other sources of floristic information.

The taxon lookup API allows users to retrieve information about a specific taxon name using the unique IPNI ID. If this is not known, it can be found out using the POWO search API.

See also

Examples

# retrieve information for a taxon name
lookup_powo("271445-2")

#> <POWO taxon id: 271445-2>
#> Name: Abbevillea chrysophylla
#> Authors: O.Berg
#> Status: Synonym
#> Rank: SPECIES
#> Accepted taxon ID: 60447743-2
#> Synonyms: 0
#> Includes distribution: FALSE

# print a summary of the returned information
r <- lookup_powo("271445-2")
print(r)

#> <POWO taxon id: 271445-2>
#> Name: Abbevillea chrysophylla
#> Authors: O.Berg
#> Status: Synonym
#> Rank: SPECIES
#> Accepted taxon ID: 60447743-2
#> Synonyms: 0
#> Includes distribution: FALSE

# format returned record into a tibble
r <- lookup_powo("271445-2")
format(r)

#> # A tibble: 1 x 23
#>   modified bibliographicCi… genus taxonomicStatus kingdom phylum family
#>   <chr>    <chr>            <chr> <chr>           <chr>   <chr>  <chr> 
#> 1 2014-03… IPNI 2021. Publ… Abbe… Synonym         Plantae Magno… Myrta…
#> # … with 16 more variables: nomenclaturalCode <chr>, source <chr>,
#> #   namePublishedInYear <int>, nomenclaturalStatus <chr>, accepted <list>,
#> #   synonym <lgl>, plantae <lgl>, fungi <lgl>, fqId <chr>, name <chr>,
#> #   authors <chr>, species <chr>, rank <chr>, reference <chr>,
#> #   classification <list>, basionymOf <list>

# format the returned list of synonyms into a tibble
r <- lookup_wcvp("60447743-2")
formatted <- format(r)
tidyr::unnest(formatted, cols=synonyms, names_sep="_")

#> # A tibble: 11 x 27
#>    id    link  name  display authors family genus specificEpithet
#>    <chr> <chr> <chr> <chr>   <chr>   <chr>  <chr> <chr>          
#>  1 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  2 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  3 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  4 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  5 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  6 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  7 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  8 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  9 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#> 10 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#> 11 6044… /tax… Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#> # … with 19 more variables: infraspecificEpithet <chr>, rank <chr>,
#> #   subordinateRank <chr>, status <chr>, publication <chr>, reviewStatus <chr>,
#> #   parent <list>, accepted <chr>, synonyms_id <chr>, synonyms_fqId <chr>,
#> #   synonyms_url <chr>, synonyms_display <chr>, synonyms_accepted <lgl>,
#> #   synonyms_family <chr>, synonyms_name <chr>, synonyms_author <chr>,
#> #   synonyms_rank <chr>, hierarchy <list>, children <list>

# format the returned list of children into a tibble
r <- lookup_wcvp("30000055-2")
formatted <- format(r)
tidyr::unnest(formatted, cols=children, names_sep="_")

#> # A tibble: 42 x 27
#>    id    link  name  display authors family genus specificEpithet
#>    <chr> <chr> <chr> <chr>   <chr>   <chr>  <chr> <chr>          
#>  1 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#>  2 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#>  3 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#>  4 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#>  5 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#>  6 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#>  7 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#>  8 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#>  9 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#> 10 3000… /tax… Camp… <b><i>… Ruiz &… Myrta… Camp… NA             
#> # … with 32 more rows, and 19 more variables: infraspecificEpithet <chr>,
#> #   rank <chr>, subordinateRank <chr>, status <chr>, publication <chr>,
#> #   reviewStatus <chr>, parent <chr>, accepted <chr>, synonyms <list>,
#> #   hierarchy <list>, children_id <chr>, children_fqId <chr>,
#> #   children_url <chr>, children_display <chr>, children_accepted <lgl>,
#> #   children_family <chr>, children_name <chr>, children_author <chr>,
#> #   children_rank <chr>