Request the next page of search results for a kewr results object.

Request the next page of search results for a kewr results object.

request_next(object)

# S3 method for wcvp_search
request_next(object)

# S3 method for powo_search
request_next(object)

# S3 method for ipni_search
request_next(object)

# S3 method for tol_search
request_next(object)

Arguments

object

A kewr search results object.

Value

The next page of results as a kewr search object of the same type.

Methods (by class)

wcvp_search: Request the next page of WCVP search results.

powo_search: Request the next page of POWO search results.

ipni_search: Request the next page of IPNI search results.

tol_search: Request the next page of ToL search results.

Examples

r <- search_wcvp("Poa")
request_next(r)
#> <WCVP search: 'Poa' filters: 'none'>
#> total results: 2461
#> returned results: 50
#> total pages: 50
#> current page: 50
#> List of 1
#>  $ :List of 9
#>   ..$ id      : chr "77076102-1"
#>   ..$ fqId    : chr "urn:lsid:ipni.org:names:77076102-1"
#>   ..$ url     : chr "/taxon/77076102-1"
#>   ..$ display : chr "<b><i>Poa alberti </i> subsp. <i>arnoldii</i> (Melderis) Olonova & G.H.Zhu</b>"
#>   ..$ accepted: logi TRUE
#>   ..$ family  : chr "Poaceae"
#>   ..$ name    : chr "Poa alberti subsp. arnoldii"
#>   ..$ author  : chr "(Melderis) Olonova & G.H.Zhu"
#>   ..$ rank    : chr "Subspecies"

# \donttest{
r <- search_powo("Poa")
request_next(r)
#> <POWO search: 'Poa' filters: 'none'>
#> total results: 2473
#> returned results: 50
#> total pages: 50
#> First result:
#> List of 1
#>  $ :List of 8
#>   ..$ accepted: logi TRUE
#>   ..$ author  : chr "Hitchc."
#>   ..$ kingdom : chr "Plantae"
#>   ..$ family  : chr "Poaceae"
#>   ..$ name    : chr "Poa albescens"
#>   ..$ rank    : chr "Species"
#>   ..$ url     : chr "/taxon/urn:lsid:ipni.org:names:283374-2"
#>   ..$ fqId    : chr "urn:lsid:ipni.org:names:283374-2"
# }

r <- search_ipni("Poa")
#> No encoding supplied: defaulting to UTF-8.
request_next(r)
#> No encoding supplied: defaulting to UTF-8.
#> <IPNI search: 'Poa', filters: 'none'>
#> total results: 2936
#> returned results: 50
#> total pages: 59
#> current page: 59
#> List of 1
#>  $ :List of 30
#>   ..$ name                 : chr "Poa sect. Ircuticae"
#>   ..$ authors              : chr "Prob. & Tzvelev"
#>   ..$ publishingAuthor     : chr "Prob. & Tzvelev"
#>   ..$ authorTeam           :List of 2
#>   ..$ rank                 : chr "sect."
#>   ..$ url                  : chr "/n/77107410-1"
#>   ..$ family               : chr "Poaceae"
#>   ..$ genus                : chr "Poa"
#>   ..$ citationType         : chr "tax. nov."
#>   ..$ hybrid               : logi FALSE
#>   ..$ hybridGenus          : logi FALSE
#>   ..$ infragenus           : chr "ircuticae"
#>   ..$ inPowo               : logi FALSE
#>   ..$ linkedPublication    :List of 15
#>   ..$ publication          : chr "Bot. Zhurn. (Moscow & Leningrad)"
#>   ..$ publicationYear      : int 2010
#>   ..$ referenceCollation   : chr "95(6): 865"
#>   ..$ publicationId        : chr "1559-2"
#>   ..$ recordType           : chr "citation"
#>   ..$ reference            : chr "Bot. Zhurn. (Moscow & Leningrad) 95(6): 865. 2010 "
#>   ..$ suppressed           : logi FALSE
#>   ..$ topCopy              : logi TRUE
#>   ..$ typeName             : chr "Poa ircutica Roshev."
#>   ..$ version              : chr "1.1"
#>   ..$ id                   : chr "77107410-1"
#>   ..$ fqId                 : chr "urn:lsid:ipni.org:names:77107410-1"
#>   ..$ hasNomenclaturalNotes: logi FALSE
#>   ..$ hasTypeData          : logi TRUE
#>   ..$ hasOriginalData      : logi FALSE
#>   ..$ hasLinks             : logi FALSE

r <- search_tol("Poa")
#> No encoding supplied: defaulting to UTF-8.
request_next(r)
#> No encoding supplied: defaulting to UTF-8.
#> <ToL search: 'Poa'>
#> total results: 182
#> returned results: 50
#> total pages: 4
#> current page: 2
#> List of 1
#>  $ :List of 20
#>   ..$ age                    : NULL
#>   ..$ collector              : chr "Bauters"
#>   ..$ collector_no           : NULL
#>   ..$ country                : NULL
#>   ..$ fasta_file_url         : chr "http://sftp.kew.org/pub/paftol/current_release/fasta/by_recovery/INSDC.ERR5033324.Cyperus_pectinatus.a353.fasta"
#>   ..$ gene_stats             :List of 2
#>   ..$ genus                  :List of 3
#>   ..$ herbcat_url            : NULL
#>   ..$ id                     : int 1111
#>   ..$ is_suspicious_placement: logi FALSE
#>   ..$ material_source        :List of 2
#>   ..$ museum_barcode         : NULL
#>   ..$ project                :List of 2
#>   ..$ raw_reads              :List of 1
#>   ..$ sequence_id            : int 2953
#>   ..$ species                :List of 2
#>   ..$ specimen_reference     : chr "KB 2015-192 (GENT)"
#>   ..$ specimen_source        : chr "Ghent University"
#>   ..$ taxonomy               :List of 4
#>   ..$ voucher_no             : chr "KB 2015-192"