Look up a name in IPNI.

Request the record for a taxon, author, or publication name in IPNI, using the IPNI ID.

lookup_ipni(id, type = c("taxon", "author", "publication"))

Arguments

id
A string containing a valid IPNI ID.

type
The type of record to look up. Either taxon, author, or publication.

Value

An ipni_taxon object, which is a simple structure with fields for each of the fields returned by the lookup API, as well as the the httr response object.

Details

The International Plant Names Index (IPNI) is a service that provides nomenclatural information for vascular plant names.

The name lookup API allows users to retrieve information for a specific taxon, author, or publication name using the unique IPNI ID. If this is not known, it can be found out using the IPNI search API.

See also

Examples

# retrieve nomenclatural information for a taxon name
name <- lookup_ipni("271445-2", "taxon")
print(name)

#> <IPNI name id: 271445-2, type: citation>
#> Name: Abbevillea chrysophylla
#> Authors: O.Berg
#> Publication: Fl. Bras. (Martius)
#> Publication Year: 1857
#> Reference: Fl. Bras. (Martius) 14(1): 437. 1857 [15 May 1857] 
#> Rank: spec.
#> In POWO: TRUE

# tidy the results in a table
tidy(name)

#> # A tibble: 1 x 39
#>   name  authors publishingAuthor authorTeam rank  url   family genus species
#>   <chr> <chr>   <chr>            <list>     <chr> <chr> <chr>  <chr> <chr>  
#> 1 Abbe… O.Berg  O.Berg           <tibble [… spec. /n/2… Myrta… Abbe… chryso…
#> # … with 30 more variables: collectionNumber <chr>, collectorTeam <chr>,
#> #   citationType <chr>, distribution <chr>, hybrid <lgl>, hybridGenus <lgl>,
#> #   inPowo <lgl>, nomenclaturalSynonym <list>, pageAsText <chr>,
#> #   linkedPublication <list>, publication <chr>, publicationYear <int>,
#> #   publicationYearNote <chr>, referenceCollation <chr>, publicationId <chr>,
#> #   recordType <chr>, reference <chr>, sameCitationAs <list>, suppressed <lgl>,
#> #   topCopy <lgl>, typeLocations <chr>, version <chr>, parent <list>, id <chr>,
#> #   fqId <chr>, hasNomenclaturalNotes <lgl>, hasTypeData <lgl>,
#> #   hasOriginalData <lgl>, hasLinks <lgl>, bhlLink <chr>

# retrieve nomenclatural information for an author
name <- lookup_ipni("20028192-1", type="author")
tidy(name)

#> # A tibble: 1 x 15
#>   alternativeNames dates examples forename fqId  id    recordType standardForm
#>   <chr>            <chr> <chr>    <chr>    <chr> <chr> <chr>      <chr>       
#> 1 ""               fl. … Schnell… Liam A.  urn:… 2002… author     Trethowan   
#> # … with 7 more variables: surname <chr>, suppressed <lgl>, taxonGroups <chr>,
#> #   url <chr>, version <chr>, summary <chr>, hasBhlLink <lgl>

# retrieve nomenclatural information for a publication
name <- lookup_ipni("987-2", type="publication")
tidy(name)

#> # A tibble: 1 x 15
#>   abbreviation bphNumber date  fqId  id    lcNumber precededBy recordType
#>   <chr>        <chr>     <chr> <chr> <chr> <chr>    <chr>      <chr>     
#> 1 Kew Bull.    513.05    Vol.… urn:… 987-2 QK1.K45  Bull. Mis… publicati…
#> # … with 7 more variables: suppressed <lgl>, title <chr>, url <chr>,
#> #   version <chr>, hasBhlLinks <lgl>, hasBhlTitleLink <lgl>,
#> #   hasBhlPageLink <lgl>