Look up a taxon in POWO.

Request the record for a taxon in Plants of the World Online (POWO) using the IPNI ID.

lookup_powo(taxonid, distribution = FALSE, .wait = 0.2)

Arguments

taxonid
A string containing a valid IPNI ID.

distribution
Include distribution in results (default FALSE).

.wait
Time to wait before making a request, to help rate limiting.

Value

A powo_taxon object, which is a simple structure with fields for each of the fields returned by the lookup API, as well as the the httr response object.

Details

Plants of the World Online (POWO) is a database of information on the world's flora. It curates information from published floras and other sources of floristic information.

The taxon lookup API allows users to retrieve information about a specific taxon name using the unique IPNI ID. If this is not known, it can be found out using the POWO search API.

See also

Examples

# retrieve information for a taxon name
lookup_powo("271445-2")

#> <POWO taxon id: 271445-2>
#> Name: Abbevillea chrysophylla
#> Authors: O.Berg
#> Status: Synonym
#> Rank: SPECIES
#> Accepted taxon ID: 60447743-2
#> Synonyms: 0
#> Includes distribution: FALSE

# print a summary of the returned information
r <- lookup_powo("271445-2")
print(r)

#> <POWO taxon id: 271445-2>
#> Name: Abbevillea chrysophylla
#> Authors: O.Berg
#> Status: Synonym
#> Rank: SPECIES
#> Accepted taxon ID: 60447743-2
#> Synonyms: 0
#> Includes distribution: FALSE

# tidy returned record into a tibble
r <- lookup_powo("271445-2")
tidy(r)

#> # A tibble: 1 × 23
#>   modified   bibliographicCitation  genus  taxonomicStatus kingdom phylum family
#>   <chr>      <chr>                  <chr>  <chr>           <chr>   <chr>  <chr> 
#> 1 2014-03-0… IPNI 2021. Published … Abbev… Synonym         Plantae Magno… Myrta…
#> # … with 16 more variables: nomenclaturalCode <chr>, source <chr>,
#> #   namePublishedInYear <int>, nomenclaturalStatus <chr>, synonym <lgl>,
#> #   plantae <lgl>, fungi <lgl>, accepted <list>, fqId <chr>, name <chr>,
#> #   authors <chr>, species <chr>, rank <chr>, reference <chr>,
#> #   classification <list>, basionymOf <list>

# tidy the returned list of synonyms into a tibble
r <- lookup_wcvp("60447743-2")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=synonyms, names_sep="_")

#> # A tibble: 11 × 27
#>    id         link              name  display authors family genus specificEpithet
#>    <chr>      <chr>             <chr> <chr>   <chr>   <chr>  <chr> <chr>          
#>  1 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  2 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  3 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  4 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  5 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  6 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  7 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  8 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#>  9 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#> 10 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#> 11 60447743-2 /taxon/60447743-2 Camp… <b><i>… (Cambe… Myrta… Camp… eugenioides    
#> # … with 19 more variables: infraspecificEpithet <lgl>, rank <chr>,
#> #   subordinateRank <chr>, status <chr>, publication <chr>, reviewStatus <chr>,
#> #   parent <list>, accepted <lgl>, synonyms_id <chr>, synonyms_fqId <chr>,
#> #   synonyms_url <chr>, synonyms_display <chr>, synonyms_accepted <lgl>,
#> #   synonyms_family <chr>, synonyms_name <chr>, synonyms_author <chr>,
#> #   synonyms_rank <chr>, hierarchy <list>, children <list>

# tidy the returned list of children into a tibble
r <- lookup_wcvp("30000055-2")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=children, names_sep="_")

#> # A tibble: 42 × 27
#>    id         link              name         display authors family genus specificEpithet
#>    <chr>      <chr>             <chr>        <chr>   <chr>   <chr>  <chr> <lgl>          
#>  1 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#>  2 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#>  3 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#>  4 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#>  5 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#>  6 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#>  7 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#>  8 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#>  9 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#> 10 30000055-2 /taxon/30000055-2 Campomanesia <b><i>… Ruiz &… Myrta… Camp… NA             
#> # … with 32 more rows, and 19 more variables: infraspecificEpithet <lgl>,
#> #   rank <chr>, subordinateRank <chr>, status <chr>, publication <chr>,
#> #   reviewStatus <chr>, parent <lgl>, accepted <lgl>, synonyms <list>,
#> #   hierarchy <list>, children_id <chr>, children_fqId <chr>,
#> #   children_url <chr>, children_display <chr>, children_accepted <lgl>,
#> #   children_family <chr>, children_name <chr>, children_author <chr>,
#> #   children_rank <chr>