Request the next page of search results for a kewr results object.

Request the next page of search results for a kewr results object.

request_next(object)

# S3 method for wcvp_search
request_next(object)

# S3 method for powo_search
request_next(object)

# S3 method for ipni_search
request_next(object)

Arguments

object
A kewr search results object.

Value

The next page of results as a kewr search object of the same type.

Methods (by class)

wcvp_search: Request the next page of WCVP search results.

powo_search: Request the next page of POWO search results.

ipni_search: Request the next page of IPNI search results.

Examples

r <- search_wcvp("Poa")
request_next(r)

#> <WCVP search: 'Poa' filters: 'none'>
#> total results: 2448
#> returned results: 50
#> total pages: 49
#> current page: 2
#> List of 1
#>  $ :List of 9
#>   ..$ id      : chr "77076102-1"
#>   ..$ fqId    : chr "urn:lsid:ipni.org:names:77076102-1"
#>   ..$ url     : chr "/taxon/77076102-1"
#>   ..$ display : chr "<b><i>Poa alberti </i> subsp. <i>arnoldii</i> (Melderis) Olonova & G.H.Zhu</b>"
#>   ..$ accepted: logi TRUE
#>   ..$ family  : chr "Poaceae"
#>   ..$ name    : chr "Poa alberti subsp. arnoldii"
#>   ..$ author  : chr "(Melderis) Olonova & G.H.Zhu"
#>   ..$ rank    : chr "Subspecies"

# \donttest{
r <- search_powo("Poa")
request_next(r)

#> <POWO search: 'Poa' filters: 'none'>
#> total results: 2456
#> returned results: 50
#> total pages: 50
#> First result:
#> List of 1
#>  $ :List of 9
#>   ..$ accepted : logi FALSE
#>   ..$ author   : chr "Buchanan"
#>   ..$ kingdom  : chr "Plantae"
#>   ..$ family   : chr "Poaceae"
#>   ..$ name     : chr "Poa albida"
#>   ..$ rank     : chr "Species"
#>   ..$ synonymOf:List of 5
#>   ..$ url      : chr "/taxon/urn:lsid:ipni.org:names:416448-1"
#>   ..$ fqId     : chr "urn:lsid:ipni.org:names:416448-1"

# }

r <- search_ipni("Poa")
request_next(r)

#> <IPNI search: 'Poa', filters: 'none'>
#> total results: 2929
#> returned results: 50
#> total pages: 59
#> current page: 2
#> List of 1
#>  $ :List of 30
#>   ..$ name                 : chr "Poa sect. Kolymenses"
#>   ..$ authors              : chr "Prob."
#>   ..$ publishingAuthor     : chr "Prob."
#>   ..$ authorTeam           :List of 1
#>   ..$ rank                 : chr "sect."
#>   ..$ url                  : chr "/n/894797-1"
#>   ..$ family               : chr "Poaceae"
#>   ..$ genus                : chr "Poa"
#>   ..$ citationType         : chr "tax. nov."
#>   ..$ hybrid               : logi FALSE
#>   ..$ hybridGenus          : logi FALSE
#>   ..$ infragenus           : chr "kolymenses"
#>   ..$ inPowo               : logi FALSE
#>   ..$ linkedPublication    :List of 15
#>   ..$ publication          : chr "Bot. Zhurn. (Moscow & Leningrad)"
#>   ..$ publicationYear      : int 1984
#>   ..$ publicationYearNote  : chr "1984"
#>   ..$ referenceCollation   : chr "69(2): 258"
#>   ..$ publicationId        : chr "1559-2"
#>   ..$ recordType           : chr "citation"
#>   ..$ reference            : chr "Bot. Zhurn. (Moscow & Leningrad) 69(2): 258. 1984 "
#>   ..$ suppressed           : logi FALSE
#>   ..$ topCopy              : logi TRUE
#>   ..$ version              : chr "1.3"
#>   ..$ id                   : chr "894797-1"
#>   ..$ fqId                 : chr "urn:lsid:ipni.org:names:894797-1"
#>   ..$ hasNomenclaturalNotes: logi FALSE
#>   ..$ hasTypeData          : logi FALSE
#>   ..$ hasOriginalData      : logi FALSE
#>   ..$ hasLinks             : logi FALSE